requirements.txt
asgiref==3.2.10
Django==3.1.2
django-datetime-widget2==0.9.5
django-jalali-date==0.3.1
django-quill-editor==0.1.22
django-summernote==0.8.11.6
djangorestframework==3.12.1
gunicorn==20.0.4
jdatetime==3.6.2
Pillow==8.0.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
